Implemented the changes requested
