+++ galasa-parent/dev.galasa.framework.api.authentication/src/main/java/dev/galasa/framework/api/authentication/AuthenticationServlet.java
@@ -21,7 +21,6 @@ import dev.galasa.framework.api.authentication.internal.OidcProvider;
 import dev.galasa.framework.api.authentication.internal.routes.AuthCallbackRoute;
 import dev.galasa.framework.api.authentication.internal.routes.AuthClientsRoute;
 import dev.galasa.framework.api.authentication.internal.routes.AuthRoute;
-import dev.galasa.framework.api.authentication.internal.routes.AuthTokensByLoginIdRoute;
 import dev.galasa.framework.api.authentication.internal.routes.AuthTokensDetailsRoute;
 import dev.galasa.framework.api.authentication.internal.routes.AuthTokensRoute;
 import dev.galasa.framework.api.common.BaseServlet;
@@ -72,7 +71,6 @@ public class AuthenticationServlet extends BaseServlet {
         addRoute(new AuthCallbackRoute(getResponseBuilder(), externalApiServerUrl));
         addRoute(new AuthTokensRoute(getResponseBuilder(), oidcProvider, dexGrpcClient, authStoreService, env));
         addRoute(new AuthTokensDetailsRoute(getResponseBuilder(), dexGrpcClient, authStoreService));
-        addRoute(new AuthTokensByLoginIdRoute(getResponseBuilder(), oidcProvider, dexGrpcClient, authStoreService,env));
 
         logger.info("Galasa Authentication API initialised");
     }
