+++ galasa-parent/dev.galasa.framework.api.openapi/src/main/resources/openapi.yaml
@@ -63,7 +63,7 @@ paths:
                 missingparametererror:
                   value:
                     error_code: 5400
-                    error_message: "E: Error occured when trying to execute request '/user'. Please check your request parameters or report the problem to your Galasa Ecosystem owner."
+                    error_message: "E: Error occurred when trying to execute request '/user'. Please check your request parameters or report the problem to your Galasa Ecosystem owner."
                   summary: One or more required query parameters are missing
         '500':
           $ref: '#/components/responses/InternalServerError'
@@ -183,6 +183,14 @@ paths:
         in the 'Authorization' header (e.g. 'Authorization: Bearer <bearer-token>').
       tags:
       - Authentication API
+      parameters:
+      - name: loginId
+        in: query
+        description: |
+            The loginId of the user whose details will be returned.
+        required: false
+        schema:
+          type: string
       responses:
         '200':
           description: Returns a JSON Object containing the list of tokens
@@ -190,6 +198,17 @@ paths:
             application/json:
               schema:
                 $ref: '#/components/schemas/AuthTokens'
+        '400':
+          description: Bad Request
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/APIError'
+              examples:
+                missingparametererror:
+                  value:
+                    error_code: 5400
+                    error_message: "E: Error occured when trying to execute request '/tokens'. Please check your request parameters or report the problem to your Galasa Ecosystem owner."        
         '401':
           $ref: "#/components/responses/Unauthorized"
         '500':
@@ -285,47 +304,6 @@ paths:
                     error_message: "GAL5000E: Error occurred when trying to access the endpoint. Report the problem to your Galasa Ecosystem owner."
                   summary: An Error occurred when trying to access the endpoint
 
-  /auth/getTokensByLoginId:
-    parameters:
-      - $ref: '#/components/parameters/ClientApiVersion'
-    get:
-      operationId: getTokensByLoginId
-      summary: Gets all access tokens of a user based on the loginId provided
-      description: |
-        Returns an array of access token objects of a user based on their loginId.
-      tags:
-      - Authentication API
-      parameters:
-      - name: loginId
-        in: query
-        description: |
-            The loginId of the user whose details will be returned.
-        required: true
-        schema:
-          type: string
-      responses:
-        '200':
-          description: Success message returning an array of access tokens.
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/AuthTokens'
-        '400':
-          description: Bad Request
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/APIError'
-              examples:
-                missingparametererror:
-                  value:
-                    error_code: 5400
-                    error_message: "E: Error occured when trying to execute request '/getTokensByLoginId'. Please check your request parameters or report the problem to your Galasa Ecosystem owner."
-                  summary: One or more required query parameters are missing      
-        '401':
-          $ref: "#/components/responses/Unauthorized"
-        '500':
-          $ref: '#/components/responses/InternalServerError'
                   
 ##################################################################################
 # Users API
